jumpy ball
You have a jumpy ball. You throw it up from the ground upto a height of
H. The ball is jumpy, so when it comes back to the ground, it rebounds
and goes up again to a distance of floor(H/2). This keeps on repeating
till it can't rebound anymore. You need to find the total distance
travelled by the ball.